added ability to have a invisible class
If you want to use a class that is the parent of the OnVisible in a transition effect you need to be able to have a class for before it was visible otherwise the animation won't override the before you set on the class.

for example:

<section className="fade-in">
	<OnVisible>
		<Article {...props} />
	</OnVisible>
</section>

.fade-in {
  div.visible {
    @include fade-in(3s);
  }
  div {
    opacity: 0;
  }
}

if I were to use this I would NOT see the item after as the opacity 0; would still be the dominat style over the fade-in (it's an animation that fades from 0-1 and uses animation-fill-mode: forwards; to change the visibilty to 1 at the end.

however if I have .invisible to me I can do the following:

.fade-in {
  div.visible {
    @include fade-in(3s);
  }
  div.invisible {
    opacity: 0;
  }
}

which is great because it works for me!