don't ever set holder to null
For what ever reason, during a page change I was finding that the ref was becoming null for no reason. This PR makes sure that if the holder element was good, then during a transition it goes bad, it won't over-ride with null, it will just cancel the assignment of the new handler.