changed debounce out for bindRaf
the debounce that was setup for calling on scroll/resize was 10ms. All browsers that I know of set a FPS cap of 60fps. 1s/60f = 16ms.

So for two reasons I have changed to bindRaf
1. There is no reason to check if we are below the desired point to change the element to visable until we are rendering a new frame, now if the fps tanks to say 25fps the calculations will only call once every 40ms, so 1/4 the calcs when the system is struggling.
2. setTimeout is not guarinteed to call on the time specified. See https://developer.mozilla.org/en-US/docs/Web/API/WindowOrWorkerGlobalScope/setTimeout#Reasons_for_delays_longer_than_specified

I came to the conclution that this was required after reading this SOF answer: https://stackoverflow.com/questions/41740082/scroll-events-requestanimationframe-vs-requestidlecallback-vs-passive-event-lis

Cheers.